+++ pkg/process/finders/register.go
@@ -23,13 +23,11 @@ import (
 	"github.com/apache/skywalking-rover/pkg/process/finders/kubernetes"
 
 	"github.com/apache/skywalking-rover/pkg/process/finders/base"
-	"github.com/apache/skywalking-rover/pkg/process/finders/scanner"
 )
 
 var finders = make(map[reflect.Type]base.ProcessFinder)
 
 func init() {
-	registerFinder(reflect.TypeOf(&scanner.Config{}), &scanner.ProcessFinder{})
 	registerFinder(reflect.TypeOf(&kubernetes.Config{}), &kubernetes.ProcessFinder{})
 }